refactor: move view logic out of PurchaseConfirmationScreen
