+++ src/stacks-hierarchy/Root_PurchaseConfirmationScreen/Root_PurchaseConfirmationScreen.tsx
@@ -1,51 +1,30 @@
+import {useAnalytics} from '@atb/analytics';
 import {MasterCard, Vipps, Visa} from '@atb/assets/svg/color/icons/ticketing';
 import {useBottomSheet} from '@atb/components/bottom-sheet';
 import {Button} from '@atb/components/button';
 import {MessageInfoBox} from '@atb/components/message-info-box';
+import {PressableOpacity} from '@atb/components/pressable-opacity';
 import {FullScreenHeader} from '@atb/components/screen-header';
 import {ThemeText} from '@atb/components/text';
-import {
-  getReferenceDataName,
-  TariffZone,
-  useFirestoreConfiguration,
-} from '@atb/configuration';
+import {useOtherDeviceIsInspectableWarning} from '@atb/fare-contracts/utils';
+import {GlobalMessage, GlobalMessageContextEnum} from '@atb/global-messages';
+import {RootStackScreenProps} from '@atb/stacks-hierarchy/navigation-types';
 import {StyleSheet, useTheme} from '@atb/theme';
 import {PaymentType, ReserveOffer} from '@atb/ticketing';
 import {
-  dictionary,
-  getTextForLanguage,
-  Language,
   PurchaseConfirmationTexts,
+  dictionary,
   useTranslation,
 } from '@atb/translations';
-import {formatToLongDateTime, secondsToDuration} from '@atb/utils/date';
-import {formatDecimalNumber} from '@atb/utils/numbers';
 import {addMinutes} from 'date-fns';
 import React, {useMemo, useState} from 'react';
-import {
-  ActivityIndicator,
-  ScrollView,
-  StyleProp,
-  View,
-  ViewStyle,
-} from 'react-native';
-import {useOtherDeviceIsInspectableWarning} from '@atb/fare-contracts/utils';
-import {
-  useOfferState,
-  UserProfileWithCountAndOffer,
-} from '../Root_PurchaseOverviewScreen/use-offer-state';
-import {SelectPaymentMethodSheet} from './components/SelectPaymentMethodSheet';
+import {ActivityIndicator, ScrollView, View} from 'react-native';
+import {useOfferState} from '../Root_PurchaseOverviewScreen/use-offer-state';
 import {usePreviousPaymentMethods} from '../saved-payment-utils';
 import {PaymentMethod} from '../types';
-import {RootStackScreenProps} from '@atb/stacks-hierarchy/navigation-types';
-import {GenericSectionItem, Section} from '@atb/components/sections';
-import {useAnalytics} from '@atb/analytics';
-import {StopPlaceFragment} from '@atb/api/types/generated/fragments/stop-places';
-import {GlobalMessage, GlobalMessageContextEnum} from '@atb/global-messages';
-import {useShowValidTimeInfoEnabled} from '../Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_TripSearchScreen/use-show-valid-time-info-enabled';
-import {PressableOpacity} from '@atb/components/pressable-opacity';
-import {MessageInfoText} from '@atb/components/message-info-text';
-import {formatPhoneNumber} from '@atb/utils/phone-number-utils.ts';
+import {PreassignedFareContractSummary} from './PreassignedFareProductSummary';
+import {SelectPaymentMethodSheet} from './components/SelectPaymentMethodSheet';
+import {ZoneSelectionMode} from '@atb-as/config-specs';
 
 type Props = RootStackScreenProps<'Root_PurchaseConfirmationScreen'>;
 
@@ -55,12 +34,10 @@ export const Root_PurchaseConfirmationScreen: React.FC<Props> = ({
 }) => {
   const styles = useStyles();
   const {theme} = useTheme();
-  const {t, language} = useTranslation();
+  const {t} = useTranslation();
   const {open: openBottomSheet, close: closeBottomSheet} = useBottomSheet();
-  const {vatPercent} = useFirestoreConfiguration();
   const {previousPaymentMethod, recurringPaymentMethods} =
     usePreviousPaymentMethods();
-  const isShowValidTimeInfoEnabled = useShowValidTimeInfoEnabled();
   const analytics = useAnalytics();
 
   const inspectableTokenWarningText = useOtherDeviceIsInspectableWarning();
@@ -80,16 +57,9 @@ export const Root_PurchaseConfirmationScreen: React.FC<Props> = ({
     recipient,
   } = params;
 
-  const {travellerSelectionMode, zoneSelectionMode} =
-    fareProductTypeConfig.configuration;
-
-  const offerEndpoint =
-    zoneSelectionMode === 'none'
-      ? 'authority'
-      : zoneSelectionMode === 'multiple-stop-harbor'
-      ? 'stop-places'
-      : 'zones';
-
+  const offerEndpoint = getOfferEndpoint(
+    fareProductTypeConfig.configuration.zoneSelectionMode,
+  );
   const isOnBehalfOf = !!recipient;
 
   const preassignedFareProductAlternatives = useMemo(
@@ -115,36 +85,19 @@ export const Root_PurchaseConfirmationScreen: React.FC<Props> = ({
     travelDate,
   );
 
-  const offerExpirationTime =
-    offerSearchTime && addMinutes(offerSearchTime, 30).getTime();
-
   const offers: ReserveOffer[] = userProfilesWithCountAndOffer.map(
     ({count, offer: {offer_id}}) => ({
       count,
       offer_id,
     }),
   );
-  const fromPlaceName = getPlaceName(fromPlace, language);
-
-  const toPlaceName = getPlaceName(toPlace, language);
-  const vatAmount = totalPrice - totalPrice / (1 + vatPercent / 100);
-
-  const vatAmountString = formatDecimalNumber(vatAmount, language);
-  const vatPercentString = formatDecimalNumber(vatPercent, language);
-  const totalPriceString = formatDecimalNumber(totalPrice, language, 2);
-
-  const travelDateText = travelDate
-    ? t(
-        PurchaseConfirmationTexts.travelDate.futureDate(
-          formatToLongDateTime(travelDate, language),
-        ),
-      )
-    : t(PurchaseConfirmationTexts.travelDate.now);
 
   function goToPayment(
     method: PaymentMethod,
     shouldSavePaymentMethod: boolean,
   ) {
+    const offerExpirationTime =
+      offerSearchTime && addMinutes(offerSearchTime, 30).getTime();
     if (offerExpirationTime && totalPrice > 0) {
       if (offerExpirationTime < Date.now()) {
         refreshOffer();
@@ -202,54 +155,6 @@ export const Root_PurchaseConfirmationScreen: React.FC<Props> = ({
       );
     });
   }
-  function summary(text?: string) {
-    // Don't render the text if the text is undefined or empty.
-    if (!text) return null;
-
-    return (
-      <ThemeText
-        style={styles.smallTopMargin}
-        type="body__secondary"
-        color="secondary"
-        testID="summaryText"
-      >
-        {text}
-      </ThemeText>
-    );
-  }
-
-  const SummaryText = () => {
-    switch (zoneSelectionMode) {
-      case 'multiple-stop-harbor':
-        return summary(
-          t(
-            PurchaseConfirmationTexts.validityTexts.harbor.messageInHarborZones,
-          ),
-        );
-      case 'none':
-        return summary(
-          getTextForLanguage(
-            preassignedFareProduct.description ?? [],
-            language,
-          ),
-        );
-      default:
-        return summary(
-          fromPlace.id === toPlace.id
-            ? t(
-                PurchaseConfirmationTexts.validityTexts.zone.single(
-                  fromPlaceName,
-                ),
-              )
-            : t(
-                PurchaseConfirmationTexts.validityTexts.zone.multiple(
-                  fromPlaceName,
-                  toPlaceName,
-                ),
-              ),
-        );
-    }
-  };
 
   return (
     <View style={styles.container}>
@@ -259,131 +164,30 @@ export const Root_PurchaseConfirmationScreen: React.FC<Props> = ({
         globalMessageContext={GlobalMessageContextEnum.appTicketing}
       />
       <ScrollView style={styles.infoSection}>
-        <View>
-          {error && (
-            <MessageInfoBox
-              type="error"
-              title={t(PurchaseConfirmationTexts.errorMessageBox.title)}
-              message={t(PurchaseConfirmationTexts.errorMessageBox.message)}
-              onPressConfig={{
-                action: refreshOffer,
-                text: t(dictionary.retry),
-              }}
-              style={styles.errorMessage}
-            />
-          )}
-          <Section>
-            <GenericSectionItem>
-              <View accessible={true} style={styles.ticketInfoContainer}>
-                <ThemeText>
-                  {getReferenceDataName(preassignedFareProduct, language)}
-                </ThemeText>
-                {recipient && (
-                  <ThemeText
-                    type="body__secondary"
-                    color="secondary"
-                    style={styles.sendingToText}
-                    testID="onBehalfOfText"
-                  >
-                    {t(
-                      PurchaseConfirmationTexts.sendingTo(
-                        recipient.name ||
-                          formatPhoneNumber(recipient.phoneNumber),
-                      ),
-                    )}
-                  </ThemeText>
-                )}
-                {fareProductTypeConfig.direction &&
-                  summary(
-                    t(
-                      PurchaseConfirmationTexts.validityTexts.direction[
-                        fareProductTypeConfig.direction
-                      ](fromPlaceName, toPlaceName),
-                    ),
-                  )}
-                <SummaryText />
-                {!isSearchingOffer &&
-                  validDurationSeconds &&
-                  isShowValidTimeInfoEnabled &&
-                  summary(
-                    t(
-                      PurchaseConfirmationTexts.validityTexts.time(
-                        secondsToDuration(validDurationSeconds, language),
-                      ),
-                    ),
-                  )}
-                {fareProductTypeConfig.configuration.requiresTokenOnMobile &&
-                  summary(
-                    t(
-                      PurchaseConfirmationTexts.validityTexts.harbor
-                        .onlyOnPhone,
-                    ),
-                  )}
-                <GlobalMessage
-                  style={styles.globalMessage}
-                  globalMessageContext={
-                    GlobalMessageContextEnum.appPurchaseConfirmation
-                  }
-                  textColor="secondary"
-                  ruleVariables={{
-                    preassignedFareProductType: preassignedFareProduct.type,
-                  }}
-                />
-                {!fareProductTypeConfig.isCollectionOfAccesses && (
-                  <MessageInfoText
-                    style={styles.smallTopMargin}
-                    type="info"
-                    message={travelDateText}
-                    textColor="secondary"
-                  />
-                )}
-              </View>
-            </GenericSectionItem>
-          </Section>
-        </View>
-
-        <Section style={styles.paymentSummaryContainer}>
-          {travellerSelectionMode !== 'none' && (
-            <GenericSectionItem>
-              {userProfilesWithCountAndOffer.map((u, i) => (
-                <PricePerUserProfile
-                  key={u.id}
-                  userProfile={u}
-                  style={i != 0 ? styles.smallTopMargin : undefined}
-                />
-              ))}
-            </GenericSectionItem>
-          )}
-          <GenericSectionItem>
-            <View style={styles.totalPaymentContainer} accessible={true}>
-              <View style={styles.totalContainerHeadings}>
-                <ThemeText type="body__primary">
-                  {t(PurchaseConfirmationTexts.totalCost.title)}
-                </ThemeText>
-                <ThemeText type="body__tertiary" color="secondary">
-                  {t(
-                    PurchaseConfirmationTexts.totalCost.label(
-                      vatPercentString,
-                      vatAmountString,
-                    ),
-                  )}
-                </ThemeText>
-              </View>
-
-              {!isSearchingOffer ? (
-                <ThemeText type="body__primary--jumbo" testID="totalPrice">
-                  {totalPriceString} kr
-                </ThemeText>
-              ) : (
-                <ActivityIndicator
-                  size={theme.spacings.medium}
-                  color={theme.text.colors.primary}
-                  style={{margin: theme.spacings.medium}}
-                />
-              )}
-            </View>
-          </GenericSectionItem>
-        </Section>
+        {error && (
+          <MessageInfoBox
+            type="error"
+            title={t(PurchaseConfirmationTexts.errorMessageBox.title)}
+            message={t(PurchaseConfirmationTexts.errorMessageBox.message)}
+            onPressConfig={{
+              action: refreshOffer,
+              text: t(dictionary.retry),
+            }}
+            style={styles.errorMessage}
+          />
+        )}
+        <PreassignedFareContractSummary
+          fareProductTypeConfig={fareProductTypeConfig}
+          fromPlace={fromPlace}
+          toPlace={toPlace}
+          isSearchingOffer={isSearchingOffer}
+          preassignedFareProduct={preassignedFareProduct}
+          totalPrice={totalPrice}
+          userProfilesWithCountAndOffer={userProfilesWithCountAndOffer}
+          recipient={recipient}
+          travelDate={travelDate}
+          validDurationSeconds={validDurationSeconds}
+        />
         {inspectableTokenWarningText && !recipient && (
           <MessageInfoBox
             type="warning"
@@ -480,81 +284,6 @@ export const Root_PurchaseConfirmationScreen: React.FC<Props> = ({
   );
 };
 
-const PricePerUserProfile = ({
-  userProfile,
-  style,
-}: {
-  userProfile: UserProfileWithCountAndOffer;
-  style: StyleProp<ViewStyle>;
-}) => {
-  const styles = useStyles();
-  const {t, language} = useTranslation();
-  const {count, offer} = userProfile;
-
-  const price = count * (offer.prices[0].amount_float || 0);
-  const originalPrice = count * (offer.prices[0].original_amount_float || 0);
-
-  const priceString = formatDecimalNumber(price, language, 2);
-  const originalPriceString = originalPrice
-    ? formatDecimalNumber(originalPrice, language, 2)
-    : undefined;
-
-  const hasFlexDiscount = price < originalPrice;
-
-  const userProfileName = getReferenceDataName(userProfile, language);
-  const a11yLabel = [
-    `${count} ${userProfileName}`,
-    `${priceString} kr`,
-    `${
-      hasFlexDiscount
-        ? `${t(
-            PurchaseConfirmationTexts.ordinaryPricePrefixA11yLabel,
-          )} ${originalPriceString} kr`
-        : ''
-    }`,
-  ].join(',');
-
-  return (
-    <View
-      accessible={true}
-      accessibilityLabel={a11yLabel}
-      style={[style, styles.userProfileItem]}
-    >
-      <ThemeText
-        style={styles.userProfileCountAndName}
-        color="secondary"
-        type="body__secondary"
-        testID="userProfileCountAndName"
-      >
-        {count} {userProfileName}
-      </ThemeText>
-      <View style={styles.userProfilePrice}>
-        {hasFlexDiscount && (
-          <ThemeText
-            type="body__tertiary"
-            color="secondary"
-            style={styles.userProfileOriginalPriceAmount}
-          >
-            {originalPriceString} kr
-          </ThemeText>
-        )}
-        <ThemeText color="secondary" type="body__secondary">
-          {priceString} kr
-        </ThemeText>
-      </View>
-    </View>
-  );
-};
-
-function getPlaceName(
-  place: TariffZone | StopPlaceFragment,
-  language: Language,
-) {
-  return 'geometry' in place
-    ? getReferenceDataName(place, language)
-    : place.name;
-}
-
 function getPaymentTypeSvg(paymentType: PaymentType) {
   switch (paymentType) {
     case PaymentType.Mastercard:
@@ -563,6 +292,17 @@ function getPaymentTypeSvg(paymentType: PaymentType) {
       return Vipps;
     case PaymentType.Visa:
       return Visa;
+  }
+}
+
+function getOfferEndpoint(zoneSelectionMode: ZoneSelectionMode) {
+  switch (zoneSelectionMode) {
+    case 'none':
+      return 'authority';
+    case 'multiple-stop-harbor':
+      return 'stop-places';
+    default:
+      return 'zones';
   }
 }
 
@@ -590,44 +330,7 @@ const useStyles = StyleSheet.createThemeHook((theme) => ({
     marginBottom: theme.spacings.medium,
   },
   infoSection: {padding: theme.spacings.medium},
-  userProfileItem: {
-    flex: 1,
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    justifyContent: 'space-between',
-  },
-  userProfileCountAndName: {marginRight: theme.spacings.small},
-  userProfilePrice: {flexDirection: 'row', flexWrap: 'wrap'},
-  userProfileOriginalPriceAmount: {
-    marginEnd: theme.spacings.small,
-    alignSelf: 'flex-end',
-    textDecorationLine: 'line-through',
-  },
-  paymentSummaryContainer: {
-    marginVertical: theme.spacings.medium,
-  },
-  sendingToText: {
-    marginTop: theme.spacings.xSmall,
-  },
-  ticketInfoContainer: {
-    flex: 1,
-  },
-  totalPaymentContainer: {
-    flexDirection: 'row',
-    width: '100%',
-    flexWrap: 'wrap',
-    justifyContent: 'space-between',
-  },
-  totalContainerHeadings: {
-    paddingVertical: theme.spacings.xSmall,
-  },
-  globalMessage: {
-    marginTop: theme.spacings.small,
-  },
   purchaseInformation: {
     marginBottom: theme.spacings.medium,
-  },
-  smallTopMargin: {
-    marginTop: theme.spacings.xSmall,
   },
 }));
